Refactor: IdResponse 패키지 변경
